voice-agent-day6/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
assemblyai
python-dotenv
